[bindings/python] Remove header polymorphic-convertible.hpp
