+++ bindings/python/src/expose-costs.cpp
@@ -5,7 +5,7 @@
 
 #include "aligator/modelling/costs/quad-costs.hpp"
 #include "aligator/modelling/costs/constant-cost.hpp"
-#include "aligator/python/polymorphic-convertible.hpp"
+#include "aligator/python/fwd.hpp"
 
 namespace aligator {
 namespace python {
