+++ bindings/python/src/expose-stage.cpp
@@ -1,6 +1,6 @@
 /// @file
 /// @copyright Copyright (C) 2022-2024 LAAS-CNRS, INRIA
-#include "aligator/python/polymorphic-convertible.hpp"
+#include "aligator/python/fwd.hpp"
 #include "aligator/python/visitors.hpp"
 
 #include "aligator/core/stage-model.hpp"
