+++ bindings/python/src/modelling/expose-explicit-dynamics.cpp
@@ -1,6 +1,6 @@
 /// @file
 /// @copyright Copyright (C) 2023 LAAS-CNRS, INRIA
-#include "aligator/python/polymorphic-convertible.hpp"
+#include "aligator/python/fwd.hpp"
 #include "aligator/python/visitors.hpp"
 #include "aligator/python/modelling/explicit-dynamics.hpp"
 #include "aligator/modelling/linear-discrete-dynamics.hpp"
