+++ bindings/python/src/modelling/expose-explicit-integrators.cpp
@@ -1,8 +1,8 @@
 /// @file
 /// @copyright Copyright (C) 2022-2024 LAAS-CNRS, INRIA
 
+#include "aligator/python/fwd.hpp"
 #include "aligator/python/modelling/explicit-dynamics.hpp"
-#include "aligator/python/polymorphic-convertible.hpp"
 
 #include "aligator/modelling/dynamics/context.hpp"
 #include "aligator/modelling/dynamics/integrator-euler.hpp"
