+++ bindings/python/src/modelling/expose-force-cost.cpp
@@ -8,7 +8,6 @@
 #include "aligator/modelling/multibody/contact-force.hpp"
 #include "aligator/modelling/multibody/multibody-wrench-cone.hpp"
 #include "aligator/modelling/multibody/multibody-friction-cone.hpp"
-#include "aligator/python/polymorphic-convertible.hpp"
 
 namespace aligator {
 namespace python {
